pipeline-manager: export some types and methods for external runners
* Splits out the pipeline-automaton out of the local runner code so that
  it can be reused.
* Exposes some existing types and adds a new async-trait to write
  different runners.
* Makes the local runner no longer prepare and use metadata.json when
  bringing up pipelines, as it is unused on the pipeline side.
