+++ crates/pipeline_manager/src/lib.rs
@@ -1,5 +1,4 @@
 mod auth;
-mod db_notifier;
 mod error;
 #[cfg(test)]
 #[cfg(feature = "integration-test")]
@@ -9,6 +8,8 @@ pub mod api;
 pub mod compiler;
 pub mod config;
 pub mod db;
+pub mod db_notifier;
 pub mod local_runner;
 pub mod logging;
+pub mod pipeline_automata;
 pub mod runner;
